Move home, global style, and site file to a site module
We also rename the `site` module to `domain` to make its purpose clearer.
