+++ WebRandomizer/ClientApp/src/site/SmHome.jsx
