+++ WebRandomizer/ClientApp/src/site/Smz3Home.jsx
