+++ WebRandomizer/ClientApp/src/site/domain.js
